requirements.txt
streamlit
pydub
pyyaml
watchdog
